Revert "upgrade to the latest 5.15-SNAPSHOT"
This reverts commit 1cae35bbbc6f0f15b35489b958f1e4186a9bed55, reversing
changes made to c244c93ebc54fc553cfeedd91ca0faedc4246245.

+++ modules/activiti-engine/src/main/java/org/activiti/engine/impl/persistence/entity/VariableScopeImpl.java
@@ -41,7 +41,6 @@ public abstract class VariableScopeImpl implements Serializable, VariableScope {
   private static final long serialVersionUID = 1L;
   
   protected Map<String, VariableInstanceEntity> variableInstances = null;
-  protected List<VariableInstanceEntity> variableInstanceList = null;
   
   protected ELContext cachedElContext;
 
@@ -54,8 +53,6 @@ public abstract class VariableScopeImpl implements Serializable, VariableScope {
   protected void ensureVariableInstancesInitialized() {
     if (variableInstances==null) {
       variableInstances = new HashMap<String, VariableInstanceEntity>();
-      variableInstanceList = new ArrayList<VariableInstanceEntity>();
-      
       CommandContext commandContext = Context.getCommandContext();
       if (commandContext == null) {
         throw new ActivitiException("lazy loading outside command context");
@@ -63,7 +60,6 @@ public abstract class VariableScopeImpl implements Serializable, VariableScope {
       List<VariableInstanceEntity> variableInstancesList = loadVariableInstances();
       for (VariableInstanceEntity variableInstance : variableInstancesList) {
         variableInstances.put(variableInstance.getName(), variableInstance);
-        variableInstanceList.add(variableInstance);
       }
     }
   }
@@ -209,15 +205,18 @@ public abstract class VariableScopeImpl implements Serializable, VariableScope {
   }
   
   public void deleteVariablesInstanceForLeavingScope() {
-    ensureVariableInstancesInitialized();
-    
-    for (VariableInstanceEntity variableInstance: variableInstanceList) {
+    List<String> variableNames = new ArrayList<String>(getVariableNamesLocal());
+    for (String variableName: variableNames) {
+      ensureVariableInstancesInitialized();
+      VariableInstanceEntity variableInstance = variableInstances.remove(variableName);
+      if (variableInstance != null) {
+        
         Context.getCommandContext().getHistoryManager()
           .recordVariableUpdate(variableInstance);
         
         variableInstance.delete();
+      }
     }
-    variableInstanceList.clear();
   }
   
   public void removeVariables(Collection<String> variableNames) {
@@ -323,7 +322,6 @@ public abstract class VariableScopeImpl implements Serializable, VariableScope {
     VariableInstanceEntity variableInstance = variableInstances.remove(variableName);
     if (variableInstance != null) {
       deleteVariableInstanceForExplicitUserCall(variableInstance, sourceActivityExecution);
-      variableInstanceList.remove(variableInstance);
     }
   }
 
@@ -351,10 +349,8 @@ public abstract class VariableScopeImpl implements Serializable, VariableScope {
 		    variableInstance.setValue(null);
 		    variableInstance.setType(newType);
 		    variableInstance.forceUpdate();
-		    variableInstance.setValue(value);
-		    VariableInstanceEntity.touch(variableInstance);
-	  } else
-		  variableInstance.setValue(value);
+	  }
+    variableInstance.setValue(value);
 
     Context.getCommandContext().getHistoryManager()
       .recordHistoricDetailVariableCreate(variableInstance, sourceActivityExecution, isActivityIdUsedForDetails());
@@ -373,7 +369,6 @@ public abstract class VariableScopeImpl implements Serializable, VariableScope {
     VariableInstanceEntity variableInstance = VariableInstanceEntity.createAndInsert(variableName, type, value);
     initializeVariableInstanceBackPointer(variableInstance);
     variableInstances.put(variableName, variableInstance);
-    variableInstanceList.add(variableInstance);
     
     // Record historic variable
     Context.getCommandContext().getHistoryManager()
